refactor: move useDebounce and useLibraryShows core into shared api-client
Eliminates duplication across web and mobile:
- useDebounce (was copy-pasted in 3 places) now lives in @showtracker/api-client
- useLibraryShows core (useInfiniteQuery + search + page flattening) is now
  shared; each platform keeps a thin wrapper that adds its scroll mechanism:
  web  → useInfiniteScroll returns an IntersectionObserver ref
  mobile → onEndReached callback for FlatList
- Removes apps/web/src/hooks/use-debounce.ts (no longer needed)
- Updates api-client package.json with react and @showtracker/shared peer deps
